+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [this versioning scheme](https://gist.github.com/cech12/69319028e88c50349a6b044000a6607b).
 
+## [1.21.7-4.5.0.0] - 2026-04-18
+### Changed
+- Updated to Minecraft 1.21.7 (Fabric 0.129.0+1.21.7, NeoForge 21.7.25-beta)
+- Updated Cloth Config support (19.0.147) (Fabric/Quilt)
+- Updated ModMenu support (15.0.2) (Fabric/Quilt)
+- Updated JEI support (23.1.0.8) (NeoForge)
+- Updated REI support (20.0.811) (NeoForge, Fabric/Quilt)
+
 ## [1.21.5-4.4.0.1] - 2026-04-18
 ### Fixed
 - fixed cake recipe to use the minecraft:eggs item tag instead of the vanilla egg item